# swagger/swaggerfile.json
{
    "info": {
        "version": "0.0.1", 
        "title": "CrowdSauce"
    }, 
    "paths": {
...
        "/api/posts/": {
            "put": {
                "operationId": "update post",
                "tags": [ "posts" ],
                "responses": {
                    "default": {"description": "Error"},
                    "200": {"description": "Success"},
                    "401": {"description": "Unauthorized"}
                },
                "parameters": [
                    {
                        "required": true, 
                        "type": "number", 
                        "description": "The facebook userId of the user", 
                        "in": "header", 
                        "name": "userid"
                    }, 
                    {
                        "required": false, 
                        "type": "string", 
                        "description": "The facebook access token of the user", 
                        "in": "header", 
                        "name": "accesstoken"
                    }, 
                    {
                        "required": true, 
                        "type": "string", 
                        "description": "The post to update", 
                        "in": "query", 
                        "name": "postId"
                    }, 
                    {
                        "required": true, 
                        "in": "body", 
                        "description": "Post data", 
                        "name": "body", 
                        "schema": {
                            "$ref": "#/definitions/Post"
                        }
                    }
                ], 
                "description": "Update a post"
            },
            "post": {
                "operationId": "create post",
                "tags": [ "posts" ],
                "responses": {
                    "default": {"description": "Error"},
                    "200": {"description": "Success"},
                    "401": {"description": "Unauthorized"}
                },
                "parameters": [
                    {
                        "required": true, 
                        "type": "integer",
                        "description": "The facebook userId of the user", 
                        "in": "header", 
                        "name": "userid"
                    }, 
                    {
                        "required": false, 
                        "type": "string", 
                        "description": "The facebook access token of the user", 
                        "in": "header", 
                        "name": "accesstoken"
                    }, 
                    {
                        "required": true, 
                        "in": "body", 
                        "description": "Post data", 
                        "name": "body", 
                        "schema": {
                            "$ref": "#/definitions/Post"
                        }
                    }
                ], 
                "description": "Create a new post"
            }, 
            "get": {
                "operationId": "get posts",
                "tags": [ "posts" ],
                "responses": {
                    "default": {"description": "Error"},
                    "200": {
                        "description": "Success", 
                        "schema": {
                            "items": {
                                "$ref": "#/definitions/Post"
                            }, 
                            "type": "array"
                        }
                    }, 
                    "401": {"description": "Unauthorized"}
                }, 
                "parameters": [
                    {
                        "required": false, 
                        "type": "number", 
                        "description": "The facebook userId of the user", 
                        "in": "header", 
                        "name": "userid"
                    }, 
                    {
                        "required": false, 
                        "type": "string", 
                        "description": "The facebook access token of the user", 
                        "in": "header", 
                        "name": "accesstoken"
                    },
                    {
                        "required": false, 
                        "type": "string", 
                        "description": "userId to query the api on", 
                        "in": "query", 
                        "name": "userId"
                    }
                ], 
                "description": "Returns a list of all the posts of a given user"
            }, 
            "delete": {
                "operationId": "delete post",
                "tags": [ "posts" ],
                "responses": {
                    "default": {"description": "Error"},
                    "200": {"description": "Success"},
                    "401": {"description": "Unauthorized"}
                },
                "parameters": [
                    {
                        "required": true, 
                        "type": "string", 
                        "description": "The facebook userId of the user", 
                        "in": "header", 
                        "name": "userid"
                    }, 
                    {
                        "required": false, 
                        "type": "string", 
                        "description": "The facebook access token of the user", 
                        "in": "header", 
                        "name": "accesstoken"
                    }, 
                    {
                        "required": true, 
                        "in": "body", 
                        "name": "body", 
                        "schema": {
                            "$ref": "#/definitions/Post"
                        }
                    }
                ], 
                "description": "Delete posts"
            }
        }, 
...
        "/api/accounts/": {
            "get": {
                "operationId": "get accounts",
                "tags": [ "accounts" ],
                "responses": {
                    "default": {"description": "Error"},
                    "200": {
                        "description": "Success", 
                        "schema": {
                            "items": {
                                "$ref": "#/definitions/Account"
                            }, 
                            "type": "array"
                        }
                    }, 
                    "401": {"description": "Unauthorized"}
                }, 
                "parameters": [
                    {
                        "required": false, 
                        "type": "number", 
                        "description": "The facebook userId of the user", 
                        "in": "header", 
                        "name": "userid"
                    }, 
                    {
                        "required": false, 
                        "type": "string", 
                        "description": "The facebook access token of the user", 
                        "in": "header", 
                        "name": "accesstoken"
                    }
                ], 
                "description": "Returns a list of all the posts of a given user"
            }, 
            "post": {
                "operationId": "create account",
                "tags": [ "accounts" ],
                "responses": {
                    "default": {"description": "Error"},
                    "200": {"description": "Success"},
                    "401": {"description": "Unauthorized"}
                },
                "parameters": [
                    {
                        "required": true, 
                        "type": "integer",
                        "description": "UserID for the account to create", 
                        "in": "header", 
                        "name": "userid"
                    }, 
                    {
                        "required": false, 
                        "type": "string", 
                        "description": "The facebook access token of the user", 
                        "in": "header", 
                        "name": "accesstoken"
                    }
                ], 
                "description": "Create new account"
            }, 
            "x-swagger-router-controller": "hello_world",
            "delete": {
                "operationId": "delete account",
                "tags": [ "accounts" ],
                "responses": {
                    "default": {"description": "Error"},
                    "200": {"description": "Success"},
                    "401": {"description": "Unauthorized"}
                },
                "parameters": [
                    {
                        "required": true, 
                        "type": "integer",
                        "description": "UserID for the account to delete", 
                        "in": "header", 
                        "name": "userid"
                    }, 
                    {
                        "required": false, 
                        "type": "string", 
                        "description": "The facebook access token of the user", 
                        "in": "header", 
                        "name": "accesstoken"
                    }
                ], 
                "description": "Delete account"
            }
        },
...
        "/api/tags/post/": {
            "post": {
                "operationId": "Add an existing tag to a post",
                "tags": [ "tags" ],
                "responses": {
                    "default": {"description": "Error"},
                    "200": {
                        "description": "Success",
                        "schema": {
                            "items": {
                                "$ref": "#/definitions/TagReturn"
                            },
                            "type": "array"
                        }
                    },
                    "401": {"description": "Unauthorized"}
                },
                "parameters": [
                    {
                        "required": true,
                        "type": "integer",
                        "description": "UserID for user",
                        "in": "header",
                        "name": "userid"
                    },
                    {
                        "required": false,
                        "type": "string",
                        "description": "The facebook access token of the user",
                        "in": "header",
                        "name": "accesstoken"
                    },
                    {
                        "required": true,
                        "description": "Contains data about the tag to be posted",
                        "in": "body",
                        "name": "body",
                        "schema": {
                            "$ref": "#/definitions/AddTag"
                        }
                    }
                ],
                "description": "Get all favorites for a given user"
            },
            "get": {
                "operationId": "Add an existing tag to a post",
                "tags": [ "tags" ],
                "responses": {
                    "default": {"description": "Error"},
                    "200": {
                        "description": "Success",
                        "schema": {
                            "items": {
                                "$ref": "#/definitions/TagReturn"
                            },
                            "type": "array"
                        }
                    },
                    "401": {"description": "Unauthorized"}
                },
                "parameters": [
                    {
                        "required": false,
                        "type": "integer",
                        "description": "UserID for user",
                        "in": "header",
                        "name": "userid"
                    },
                    {
                        "required": false,
                        "type": "string",
                        "description": "The facebook access token of the user",
                        "in": "header",
                        "name": "accesstoken"
                    }
                ],
                "description": "Get all favorites for a given user"
            }
        },
...
        "/api/favorites/post/": {
            "get": {
                "operationId": "get post favorites",
                "tags": [ "favorites" ],
                "responses": {
                    "default": {"description": "Error"},
                    "200": {
                        "description": "Success",
                        "schema": {
                            "items": {
                                "$ref": "#/definitions/FavoriteReturn"
                            },
                            "type": "array"
                        }
                    },
                    "401": {"description": "Unauthorized"}
                },
                "parameters": [
                    {
                        "required": true,
                        "type": "integer",
                        "description": "UserID for user",
                        "in": "header",
                        "name": "userid"
                    },
                    {
                        "required": false,
                        "type": "string",
                        "description": "The facebook access token of the user",
                        "in": "header",
                        "name": "accesstoken"
                    },
                    {
                        "required": true,
                        "type": "string",
                        "description": "The post to get a favorites for",
                        "in": "query",
                        "name": "postId"
                    }
                ],
                "description": "Get all posts for a given user"
            }
        }
    }, 
    "schemes": [
        "http", 
        "https"
    ], 
    "produces": [
        "application/json"
    ], 
    "basePath": "/", 
    "host": "localhost:3000",
    "definitions": {
        "Post": {
            "properties": {
                "rating": {
                    "type": "number"
                }, 
                "directions": {
                    "items": {
                        "type": "string"
                    }, 
                    "type": "array"
                }, 
                "ingredients": {
                    "items": {
                        "type": "string"
                    }, 
                    "type": "array"
                }, 
                "imageLink": {
                    "type": "string"
                }, 
                "notes": {
                    "type": "string"
                }, 
                "recipeLink": {
                    "type": "string"
                }
            }
        },
        "Account": {
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "AddTag": {
            "properties": {
                "tagName": {
                    "tagName": "string"
                },
                "postId": {
                    "tagName": "string"
                }
            }
        },
...
        "TagReturn": {
            "properties": {
                "tagName": {
                    "type": "string"
                },
                "postId": {
                    "type": "integer"
                }
            }
        },
        "FavoriteReturn": {
            "properties": {
                "postId": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        }
    }, 
    "swagger": "2.0", 
    "consumes": [
        "application/json"
    ]
}